+++ src/main/java/com/besson/endfield/block/custom/ProtocolAnchorCorePortBlock.java
@@ -5,10 +5,12 @@ import com.besson.endfield.blockEntity.ModBlockEntities;
 import com.besson.endfield.blockEntity.custom.ProtocolAnchorCoreBlockEntity;
 import com.besson.endfield.blockEntity.custom.ProtocolAnchorCorePortBlockEntity;
 import net.minecraft.core.BlockPos;
+import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.RenderShape;
 import net.minecraft.world.level.block.entity.BlockEntity;
@@ -32,15 +34,29 @@ public class ProtocolAnchorCorePortBlock extends ModBlockEntityWithFacing {
 
     @Override
     public InteractionResult use(BlockState pState, Level pLevel, BlockPos pPos, Player pPlayer, InteractionHand pHand, BlockHitResult pHit) {
-        if (!pLevel.isClientSide()) {
-            BlockEntity entity = pLevel.getBlockEntity(pPos);
-            if (entity instanceof ProtocolAnchorCorePortBlockEntity entity1) {
-                ProtocolAnchorCoreBlockEntity parent = entity1.getParentBlock();
-                System.out.println("parent:" + parent);
-                if (parent != null) {
-                    NetworkHooks.openScreen((ServerPlayer) pPlayer, parent, parent::writeScreenData);
-                    return InteractionResult.SUCCESS;
-                }
+        if (pLevel.isClientSide()) return InteractionResult.CONSUME;
+
+        BlockEntity entity = pLevel.getBlockEntity(pPos);
+        if (!(entity instanceof ProtocolAnchorCorePortBlockEntity entity1)) return InteractionResult.CONSUME;
+
+        ItemStack heldItem = pPlayer.getItemInHand(pHand);
+        if (heldItem.isEmpty() && !pPlayer.isShiftKeyDown()) {
+            ProtocolAnchorCoreBlockEntity parent = entity1.getParentBlock();
+            if (parent != null) {
+                NetworkHooks.openScreen((ServerPlayer) pPlayer, parent, parent::writeScreenData);
+                return InteractionResult.SUCCESS;
+            }
+        }
+
+        if (heldItem.isEmpty() && pPlayer.isShiftKeyDown()) {
+            entity1.clearFilter();
+            if (pPlayer instanceof ServerPlayer) {
+                ((ServerPlayer) pPlayer).displayClientMessage(Component.literal("Cleared filter"), false);
+            }
+        } else {
+            entity1.setFilter(heldItem);
+            if (pPlayer instanceof ServerPlayer) {
+                ((ServerPlayer) pPlayer).displayClientMessage(Component.literal("Set filter to: " + heldItem.getHoverName().getString()), false);
             }
         }
         return InteractionResult.CONSUME;